Remove support of reset password by an administrator
It complexifies the code, and the user experience,
as we don't want to send a simple code with a long TTL,
to the user to confirm the password change.

And we don't want anymore to send links into mails, to avoid pishing,
so we don't want anymore links we long hash code as parameters.